required data
