--- data/processed/uq_parameters_emissions.csv
@@ -1,4 +1,0 @@
-constellation,subscribers_low,subscribers_baseline,subscribers_high,global_warming,global_warming_wc,ozone_depletion,mineral_depletion,freshwater_toxicity,human_toxicity,global_warming_roct,ozone_depletion_roct,mineral_depletion_roct,freshwater_toxicity_roct,human_toxicity_roct,global_warming_ait,ozone_depletion_ait,mineral_depletion_ait,freshwater_toxicity_ait,human_toxicity_ait,global_warming_propellant,ozone_depletion_propellant,mineral_depletion_propellant,freshwater_toxicity_propellant,human_toxicity_propellant,global_warming_schd,ozone_depletion_schd,mineral_depletion_schd,freshwater_toxicity_schd,human_toxicity_schd,global_warming_trans,ozone_depletion_trans,mineral_depletion_trans,freshwater_toxicity_trans,human_toxicity_trans,global_warming_campaign,ozone_depletion_campaign,mineral_depletion_campaign,freshwater_toxicity_campaign,human_toxicity_campaign,oneweb_f9,oneweb_sz,total_global_warming_em,total_ozone_depletion_em,total_mineral_depletion,total_freshwater_toxicity,total_human_toxicity,total_climate_change,total_climate_change_wc
-Starlink,2500000,3500000,4500000,630466.1352,27266165.94,6837.18,0,0,0,4113533.907,0.277478514,1783.337118,20253601.64,1.51530806,1616263.557,0.156575296,15.65466435,7701094.993,0.486234151,4744753.339,0.546874653,32.92663196,15292951.2,1.378727964,5594990.91690164,0.49949299670200004,277.75674638,36949013.693581775,2.431925559792,17220.72491,0.003568284,0.83687164,47571.20342,0.004766684,5666556.742,0.777870405,33.22600998,18683396.82,1.695861368,637303.3152000001,291812.24960000004,22383785.32201164,6839.441860148701,2143.73804231,98927629.55000177,7.512823786792,22383785.32201164,49019485.126811646
-OneWeb,500000,800000,1000000,288655.1096,12031437.19,3157.14,0,0,0,44680412.98,3.11181773,12473.4086,280703930.5,19.1361269,1616263.557,0.156575296,15.65466435,7701094.993,0.486234151,968910.1994,0.109998823,6.71625049,3114043.098,0.28140976,3223330.48355453,0.287918554652,159.77557178400002,21269740.34670757,1.399552018292,4328.603808,0.001339551,0.158493574,22931.63867,0.002580373,5666556.742,0.777870405,33.22600998,18683396.82,1.695861368,637303.3152000001,291812.24960000004,56448457.67536253,3161.585520359652,12688.939590178,331495137.39637756,23.001764570292,56448457.67536253,68191239.75576252
-Kuiper,1500000,2500000,3500000,467816.8,107643343.2,86728.6,0,0,0,11018755.48,0.745735051,2719.725687,69735037.48,4.585385379,1616263.557,0.156575296,15.65466435,7701094.993,0.486234151,4793267.48,0.223292749,34.59642811,17124098.34,1.520253794,8984275.33639199,0.7539977624649999,510.201815652,61938424.03519763,4.181106282081999,11043.18682,0.001892801,0.194786912,17341.87779,0.001653926,5666556.742,0.777870405,33.22600998,18683396.82,1.695861368,637303.3152000001,291812.24960000004,32557978.58221199,86731.25936406446,3313.599392004,175199393.54598764,12.470494900081999,32557978.58221199,139733504.98221198
